make images stay after we move to favourites page
