+++ src/index.tsx
@@ -4,6 +4,7 @@ import './index.css';
 import reportWebVitals from './reportWebVitals';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import { FavouritesContextProvider } from './hooks/useFavorites';
+import { ImagesContextProvider } from './hooks/useImages';
 import ImagesSearch from './pages/ImagesSearch';
 import ImagesFavourites from './pages/ImagesFavourites';
 import CssBaseline from '@mui/material/CssBaseline';
@@ -24,9 +25,11 @@ const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
 root.render(
   <React.StrictMode>
     <CssBaseline />
-    <FavouritesContextProvider>
-      <RouterProvider router={router} />
-    </FavouritesContextProvider>
+    <ImagesContextProvider>
+      <FavouritesContextProvider>
+        <RouterProvider router={router} />
+      </FavouritesContextProvider>
+    </ImagesContextProvider>
   </React.StrictMode>
 );
